@@ -3,34 +3,36 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@types/node": "^15.12.2",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "@types/react-router-dom": "^5.1.7",
-    "firebase": "^9.1.3",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.8",
+    "@types/node": "^17.0.8",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "firebase": "^9.6.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.8.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "recoil": "^0.3.1",
+    "react-hook-form": "^7.24.0",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
+    "recoil": "^0.5.2",
     "typeface-roboto": "^1.1.13",
-    "typescript": "^4.4.4"
+    "typescript": "^4.5.4"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^13.1.9",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/css-mediaquery": "^0.1.1",
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^27.4.0",
     "css-mediaquery": "^0.1.2",
-    "firebase-tools": "^9.20.0",
-    "husky": "^4.3.8",
-    "lint-staged": "^11.0.0",
-    "prettier": "^2.3.1",
+    "firebase-tools": "^10.1.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.1.7",
+    "prettier": "^2.5.1",
     "source-map-explorer": "^2.5.2"
   },
   "scripts": {
@@ -42,7 +44,8 @@
     "coverage": "yarn test --coverage --watchAll=false",
     "analyze": "yarn build && source-map-explorer build/static/js/main.*",
     "ci": "CI=true yarn test && CI=true yarn build",
-    "emulator:export": "yarn firebase emulators:export firestore_mock_data"
+    "emulator:export": "yarn firebase emulators:export firestore_mock_data",
+    "prepare": "husky install"
   },
   "eslintConfig": {
     "extends": [
@@ -80,13 +83,6 @@
     }
   },
   "lint-staged": {
-    "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
-      "prettier --write --single-quote --no-semi --jsx-single-quote --arrow-parens avoid --trailing-comma none"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+    "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": "prettier --write --single-quote --no-semi --jsx-single-quote --arrow-parens avoid --trailing-comma none"
   }
 }
