@@ -1,4 +1,4 @@
-import { Box, BoxProps } from '@material-ui/core'
+import { Box, BoxProps } from '@mui/material'
 
 const FabWrapper = (props: BoxProps) => {
   return <Box position='fixed' pr={3} pb={3} right={0} bottom={0} {...props} />
